+++ crates/cast/src/starknet_commands/account/create.rs
@@ -1,19 +1,19 @@
-use crate::starknet_commands::account::{prepare_account_json, write_account_to_accounts_file};
-use anyhow::{anyhow, bail, Result};
+use crate::starknet_commands::account::{
+    get_account_deployment_fee, prepare_account_json, write_account_to_accounts_file,
+};
+use anyhow::{anyhow, Result};
 use camino::Utf8PathBuf;
 use cast::helpers::constants::OZ_CLASS_HASH;
 use cast::helpers::response_structs::AccountCreateResponse;
 use cast::helpers::scarb_utils::CastConfig;
-use cast::{extract_or_generate_salt, get_chain_id, get_keystore_password, parse_number};
+use cast::{extract_or_generate_salt, parse_number};
 use clap::Args;
 use serde_json::json;
-use starknet::accounts::{AccountFactory, OpenZeppelinAccountFactory};
-use starknet::core::types::{FeeEstimate, FieldElement};
+use starknet::core::types::FieldElement;
 use starknet::core::utils::get_contract_address;
 use starknet::providers::jsonrpc::HttpTransport;
 use starknet::providers::JsonRpcClient;
-use starknet::signers::{LocalWallet, SigningKey};
-use std::fs;
+use starknet::signers::SigningKey;
 
 #[derive(Args, Debug)]
 #[command(about = "Create an account with all important secrets")]
@@ -33,10 +33,6 @@ pub struct Create {
     /// Custom open zeppelin contract class hash of declared contract
     #[clap(short, long)]
     pub class_hash: Option<String>,
-
-    /// If passed, create account from keystore and starkli account JSON file
-    #[clap(short, long)]
-    pub from_keystore: bool,
 }
 
 #[allow(clippy::too_many_arguments)]
@@ -48,17 +44,8 @@ pub async fn create(
     salt: Option<FieldElement>,
     add_profile: bool,
     class_hash: Option<String>,
-    from_keystore: bool,
-    keystore_path: Option<Utf8PathBuf>,
-    account_path: Option<Utf8PathBuf>,
 ) -> Result<AccountCreateResponse> {
-    let (account_json, max_fee) = if from_keystore {
-        let keystore_path_ = keystore_path
-            .ok_or_else(|| anyhow!("--keystore must be passed when using --from-keystore"))?;
-        let account_path_ = account_path
-            .ok_or_else(|| anyhow!("--account must be passed when using --from-keystore"))?;
-        import_from_keystore(provider, keystore_path_, account_path_).await?
-    } else {
+    let (account_json, max_fee) = {
         let salt = extract_or_generate_salt(salt);
         let class_hash = {
             let ch = match &class_hash {
@@ -110,77 +97,6 @@ pub async fn create(
     })
 }
 
-pub async fn import_from_keystore(
-    provider: &JsonRpcClient<HttpTransport>,
-    keystore_path: Utf8PathBuf,
-    account_path: Utf8PathBuf,
-) -> Result<(serde_json::Value, u64)> {
-    let account_info: serde_json::Value = serde_json::from_str(&fs::read_to_string(account_path)?)?;
-    let deployment = account_info
-        .get("deployment")
-        .ok_or_else(|| anyhow!("No deployment field in account JSON file"))?;
-
-    let deployed = match deployment.get("status").and_then(serde_json::Value::as_str) {
-        Some("deployed") => true,
-        Some("undeployed") => false,
-        _ => bail!("Unknown deployment status value"),
-    };
-
-    let salt: Option<FieldElement> = {
-        let salt = deployment.get("salt").and_then(serde_json::Value::as_str);
-        match (salt, deployed) {
-            (Some(salt), _) => Some(parse_number(salt)?),
-            (None, false) => bail!("No salt field in account JSON file"),
-            (None, true) => None,
-        }
-    };
-
-    let class_hash: FieldElement = {
-        let ch = deployment
-            .get("class_hash")
-            .and_then(serde_json::Value::as_str)
-            .ok_or_else(|| anyhow!("No class_hash field in account JSON file"))?;
-        parse_number(ch)?
-    };
-
-    let private_key = SigningKey::from_keystore(keystore_path, get_keystore_password()?.as_str())?;
-    let public_key: FieldElement = {
-        let pk = account_info
-            .get("variant")
-            .and_then(|v| v.get("public_key"))
-            .and_then(serde_json::Value::as_str)
-            .ok_or_else(|| anyhow::anyhow!("No public_key in account JSON file"))?;
-        parse_number(pk)?
-    };
-    if public_key != private_key.verifying_key().scalar() {
-        bail!("Public key mismatch");
-    }
-
-    let address: FieldElement = if let Some(salt) = salt {
-        get_contract_address(salt, class_hash, &[public_key], FieldElement::ZERO)
-    } else {
-        let address = deployment
-            .get("address")
-            .and_then(serde_json::Value::as_str)
-            .ok_or_else(|| anyhow::anyhow!("No address in account JSON file"))?;
-        parse_number(address)?
-    };
-
-    let max_fee = match salt {
-        Some(salt) => {
-            get_account_deployment_fee(&private_key, class_hash, salt, provider)
-                .await?
-                .overall_fee
-        }
-        None => 0,
-    };
-
-    let account_json =
-        prepare_account_json(&private_key, address, deployed, Some(class_hash), salt);
-
-    Ok((account_json, max_fee))
-}
-
 async fn generate_account(
     provider: &JsonRpcClient<HttpTransport>,
     salt: FieldElement,
@@ -202,29 +118,4 @@ async fn generate_account(
         .overall_fee;
 
     Ok((account_json, max_fee))
-}
-
-async fn get_account_deployment_fee(
-    private_key: &SigningKey,
-    class_hash: FieldElement,
-    salt: FieldElement,
-    provider: &JsonRpcClient<HttpTransport>,
-) -> Result<FeeEstimate> {
-    let signer = LocalWallet::from_signing_key(private_key.clone());
-    let chain_id = get_chain_id(provider).await?;
-    let factory = OpenZeppelinAccountFactory::new(class_hash, chain_id, signer, provider).await?;
-    let deployment = factory.deploy(salt);
-
-    let fee_estimate = deployment.estimate_fee().await;
-
-    if let Err(err) = &fee_estimate {
-        if err
-            .to_string()
-            .contains("StarknetErrorCode.UNDECLARED_CLASS")
-        {
-            bail!("The class {class_hash} is undeclared, try using --class-hash with a class hash that is already declared");
-        }
-    }
-
-    Ok(fee_estimate?)
 }
